+++ Integration Tests/Tests/DepotBaseTests.cs
@@ -17,10 +17,17 @@ namespace Integration_Tests.Tests;
 public class DepotBaseTests
     : DepotIntegrationTestsBase<EntityProxy, DepotProxy, DatabaseProxy.DatabaseProxy> {
 
-    protected override EntityProxy EntityFactory(string Entropy) {
-        return new EntityProxy {
-        };
+    protected override async Task<EntityProxy> EntityFactoryAsync(string Entropy) {
+        return await Task.FromResult(new EntityProxy {
+
+        });
     }
+
+   protected override EntityProxy EntityFactory(string Entropy) {
+        return new EntityProxy {
+
+        };
+   }
 
     /// <summary>
     ///     Method: <see cref="DepotBase{TDatabase, TEntity}.Update(QueryInput{TEntity, UpdateInput{TEntity}})"/> 
@@ -96,4 +103,5 @@ public class DepotBaseTests
                 updatedEntity.EntityDependencyProxy.Id
             );
     }
+
 }